+++ NEWS.md
@@ -19,6 +19,7 @@
 - new `nest_join()` function. `nest_join()` creates a list column of the matching rows. `nest_join()` + `tidyr::unnest()` is equivalent to `inner_join`  (#3570). 
 - `last_col()` is re-exported from tidyselect (#3584). 
 - hybrid version of `sum(na.rm = FALSE)` exits early when there are missing values. This considerably improves performance when there are missing values early in the vector (#3288). 
+- `mutate()` removes a column when the expression evaluates to `NULL` for all groups (#2945).
 
 # dplyr 0.7.5.9001